\name{silo-package}
\alias{silo-package}
\docType{package}

\title{
\packageTitle{silo}
}

\description{
\packageDescription{silo}
}

\details{
Package for multi-criteria selection of optimal species lists.
}

\author{
\packageAuthor{silo}

Maintainer: \packageMaintainer{silo}
}

\keyword{package}

\section{Documentation Index}{\packageIndices{silo}}

\seealso{
    \code{\link[kofnGA:kofnGA-package]{kofnGA}}
}
