% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grain.R
\name{silo}
\alias{silo}
\title{Specify the current grid}
\usage{
silo(extent, dimension, projection, resample = NULL, bands = NULL, ...)
}
\arguments{
\item{extent}{spatial extent xmin,xmax,ymin,ymax or a function call (see Details)}

\item{dimension}{physical data size,  number of columns, number of rows or a function call (see Details)}

\item{projection}{map projection (understood by PROJ)}

\item{resample}{resample algorithm (understood by GDAL warper)}

\item{bands}{which bands to read (1-based, can be repeated, non-sequential)}

\item{...}{reserved}
}
\value{

}
\description{
Create or modify the details of a grid specification.
}
\details{
The default grid specification is unspecified. This might change, but it's so that
gdalio can apply its own default based on the data in use.

We can set the grid components explicitly  with 'silo(property_name = <values>)', such as
two numbers for 'dimension', or four numbers for 'extent', but these properties can also be changed
\emph{relatively} (if they are already set) using functions \code{\link[=finer]{finer()}}, \code{\link[=coarser]{coarser()}} (dimension) and
\code{\link[=bigger]{bigger()}}, \code{\link[=wider]{wider()}}, \code{\link[=taller]{taller()}}, or \code{\link[=smaller]{smaller()}}, \code{\link[=thinner]{thinner()}}, \code{\link[=shorter]{shorter()}}.
}
